feat: added routing to the appplication and moved base to its own component. #3
