+++ pages/Home.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+import Base from '../components/Base/Base';
+
+const HomePage = () => {
+    return (
+        <Base>
+            <View>
+                <Text>Hello</Text>
+            </View>
+        </Base>
+    );
+
+};
+
+export default HomePage;
