+++ types.ts
@@ -1,0 +1,3 @@
+export interface HasChildren {
+    children: any;
+}
